moved code which work with database to service file, moved isAdmin function to middleware file, moved some constants to constant file, rewrited skip parametrs to query string
