+++ src/file/file.middleware.ts
@@ -1,6 +1,8 @@
 import multer from 'multer';
-import { Request } from 'express';
+import { Request, Response } from 'express';
 import dateFormat from 'dateformat';
+import axios from 'axios';
+import { RoleMicroseviceConstants } from '../shared/constants';
 
 const storage = multer.diskStorage({
 // @ts-ignore
@@ -17,3 +19,16 @@ const storage = multer.diskStorage({
 });
 
 export const upload = multer({ storage });
+
+export const isAdmin = async (req: Request, res: Response, next: any) => {
+  try {
+    const data = await axios.get(`http://${RoleMicroseviceConstants.HOSTNAME}:${RoleMicroseviceConstants.PORT}/user/role/${req.params.userId}`);
+    if (data.data.name !== 'admin') {
+      return res.sendStatus(403);
+    }
+    return next();
+  } catch (e) {
+    res.sendStatus(500);
+    console.error(e);
+  }
+};
