+++ src/index.ts
@@ -1,15 +1,11 @@
 import express, { Express, Request, Response } from 'express';
 import bodyParser from 'body-parser';
-import { config } from 'dotenv';
 
 import routes from './index.routes';
-import { Constants } from './shared/constants';
+import { Constants, URI } from './shared/constants';
 import { connection } from './shared/tsgoose';
 
-config({ path: './.env' });
-
 const app: Express = express();
-const uri: string = process.env.URI || '';
 
 app.use(bodyParser.json());
 app.use(routes);
@@ -23,7 +19,5 @@ app.route('*')
 app.listen(process.env.PORT, process.env.HOST, async () => {
   console.log(`Server was running at http://${Constants.HOSTNAME}:${Constants.PORT}/`);
 
-  connection(uri);
+  connection(URI);
 });
-
-
